apicliente/requirements.txt
Django~=5.1.1
django-bootstrap5~=24.3
django-bootstrap-icons~=0.8.6
django-environ~=0.11.2
